requirements.txt
datasets==2.3.2
happytransformer==2.4.1
huggingface-hub==0.8.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
nltk==3.7
numpy==1.23.1
pandas==1.4.3
PyYAML==5.4.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.96
spacy==3.4.0
tokenizers==0.12.1
torch==1.12.0+cu116
torchcontrib==0.0.2
tqdm==4.64.0
transformers==4.20.1
Unidecode==1.1.1
wandb==0.13.5
text_unidecode
iterative-stratification
optuna==3.0.3